requirements/common.txt
Flask==0.10.1
Flask-Script==0.6.6
facebook-sdk==0.4.0
